# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /strategies/{strategyId}/portfolios:
    get:
      description: Returns both REAL and PAPER portfolios for a strategy (if they exist)
      operationId: PortfolioController_getStrategyPortfolios
      parameters:
        - name: strategyId
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Portfolios retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security: &ref_6
        - JWT-auth: []
      summary: Get portfolios for a strategy
      tags: &ref_7
        - Portfolios
  /portfolios/{id}:
    get:
      description: Returns detailed information about a portfolio including strategy info and calculated equity
      operationId: PortfolioController_getPortfolioById
      parameters:
        - name: id
          required: true
          in: path
          description: Portfolio ID
          schema:
            type: string
      responses:
        '200':
          description: Portfolio retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: Portfolio not found
      security: *ref_6
      summary: Get portfolio details
      tags: *ref_7
    patch:
      description: 'Deposits or withdraws funds from a portfolio. Cannot do both in same request. For PAPER portfolios: updates database only. For REAL portfolios: may require exchange API integration.'
      operationId: PortfolioController_updatePortfolio
      parameters:
        - name: id
          required: true
          in: path
          description: Portfolio ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePortfolioDto'
      responses:
        '200':
          description: Portfolio updated successfully
        '400':
          description: Invalid request - cannot deposit and withdraw simultaneously, or insufficient balance
        '401':
          description: Unauthorized
        '404':
          description: Portfolio not found
      security: *ref_6
      summary: Update portfolio (deposit/withdrawal)
      tags: *ref_7
  /portfolios/{id}/performance:
    get:
      description: 'Returns time-series performance data for a portfolio. NOTE: Historical snapshots not yet implemented - returns initial and current values only.'
      operationId: PortfolioController_getPortfolioPerformance
      parameters:
        - name: id
          required: true
          in: path
          description: Portfolio ID
          schema:
            type: string
        - name: timeframe
          required: false
          in: query
          description: Performance timeframe
          schema:
            default: 1M
            type: string
            enum:
              - 1D
              - 1W
              - 1M
              - 3M
              - 1Y
              - ALL
      responses:
        '200':
          description: Portfolio performance retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: Portfolio not found
      security: *ref_6
      summary: Get portfolio performance over time
      tags: *ref_7
  /strategies/{strategyId}/positions:
    get:
      description: Returns paginated list of positions for a strategy. Queries from Position or PaperPosition table based on mode parameter.
      operationId: PositionController_getStrategyPositions
      parameters:
        - name: strategyId
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
        - name: mode
          required: true
          in: query
          description: Trading mode (REAL or PAPER)
          schema:
            type: string
            enum:
              - REAL
              - PAPER
        - name: symbol
          required: false
          in: query
          description: Filter by symbol
          schema:
            example: BTCUSDT
            type: string
        - name: side
          required: false
          in: query
          description: Filter by position side (LONG or SHORT)
          schema:
            example: LONG
            type: string
            enum:
              - LONG
              - SHORT
        - name: status
          required: false
          in: query
          description: Filter by status (is_active)
          schema:
            type: string
            enum:
              - OPEN
              - CLOSED
        - name: cursor
          required: false
          in: query
          description: Cursor for pagination (position ID)
          schema:
            example: clxy1234567890abcdefghijk
            type: string
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            minimum: 1
            maximum: 100
            default: 20
            example: 20
            type: number
      responses:
        '200':
          description: Positions retrieved successfully with summary
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security: &ref_8
        - JWT-auth: []
      summary: Get positions for a strategy
      tags: &ref_9
        - Positions
  /positions/{id}:
    get:
      description: Returns detailed information about a specific position (REAL or PAPER based on mode parameter)
      operationId: PositionController_getPositionById
      parameters:
        - name: id
          required: true
          in: path
          description: Position ID
          schema:
            type: string
        - name: mode
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Position retrieved successfully
        '401':
          description: Unauthorized
        '404':
          description: Position not found
      security: *ref_8
      summary: Get position details
      tags: *ref_9
  /positions/{id}/chart-data:
    get:
      description: Fetches candle data from GraphQL service for visualizing the position on a chart. Returns candles before and after position entry time.
      operationId: PositionController_getPositionChartData
      parameters:
        - name: id
          required: true
          in: path
          description: Position ID
          schema:
            type: string
        - name: mode
          required: true
          in: query
          description: Trading mode (REAL or PAPER)
          schema:
            type: string
            enum:
              - REAL
              - PAPER
        - name: candles_before
          required: false
          in: query
          description: Number of candles before entry
          schema:
            minimum: 1
            maximum: 200
            default: 50
            example: 50
            type: number
        - name: candles_after
          required: false
          in: query
          description: Number of candles after entry
          schema:
            minimum: 1
            maximum: 200
            default: 50
            example: 50
            type: number
      responses:
        '200':
          description: Candle data retrieved successfully
        '400':
          description: Failed to fetch candle data
        '401':
          description: Unauthorized
        '404':
          description: Position not found
      security: *ref_8
      summary: Get candle data for position chart
      tags: *ref_9
  /dashboard/summary:
    get:
      description: 'Returns aggregated dashboard data including total strategies, portfolios value, P&L, and top performing strategies. Note: today_pnl and performance_7d are approximations. Implement historical snapshot tracking for accurate calculations.'
      operationId: DashboardController_getDashboardSummary
      parameters: []
      responses:
        '200':
          description: Dashboard summary retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_strategies:
                    type: number
                    example: 5
                  live_strategies:
                    type: number
                    example: 2
                  total_portfolios_value:
                    type: number
                    example: 25000.5
                  today_pnl:
                    type: number
                    example: 150.25
                  open_positions:
                    type: number
                    example: 3
                  strategies_by_performance:
                    type: array
                    items:
                      type: object
                      properties:
                        strategy:
                          type: object
                        portfolio:
                          type: object
                        performance_7d:
                          type: number
                          example: 5.2
                  note:
                    type: string
        '401':
          description: Unauthorized
      security:
        - JWT-auth: []
      summary: Get dashboard summary
      tags:
        - Dashboard
info:
  title: Crypto Watcher Bot API
  description: API for crypto trading signal bot with automated order execution
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    UpdatePortfolioDto:
      type: object
      properties:
        deposit:
          type: number
          description: Amount to deposit
          example: 1000
          minimum: 0
        withdrawal:
          type: number
          description: Amount to withdraw
          example: 500
          minimum: 0